extract components to separate files and use partition to get success and failed
